requirements.txt
coverage==4.0.3
django-debug-toolbar==1.4
django-extensions==1.6.7
django-grappelli==2.10.1
django-ordered-model==1.2.1
django>=1.11.0,<1.11.1024
factory_boy==2.8.1
flake8==3.2.1
ipython==5.1.0
isort==4.2.5
psycopg2==2.6.2
python-crontab==2.1.1
python-dateutil==2.6.0
pytz==2016.4
raven==5.32.0
tox==2.3.1
